:art: Adapted partially mechanic to ES module package
