+++ packages/mechanic/app/function-loader.cjs
