+++ packages/mechanic/package.json
@@ -34,7 +34,11 @@
     "src",
     "app"
   ],
-  "module": "src/index.js",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.js",
+    "./src/commands/*": "./src/commands/*.cjs"
+  },
   "dependencies": {
     "@babel/core": "^7.14.6",
     "@babel/plugin-transform-runtime": "^7.14.5",
@@ -73,6 +77,9 @@
   "devDependencies": {
     "prettier": "^2.0.5",
     "stylelint-config-prettier": "^8.0.2"
+  },
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "gitHead": "72540367e35f7c65bc111779ceb4c59ea0e1a5fb"
 }
