+++ packages/mechanic/src/commands/utils.cjs
